requirements.txt
gradio
python-dotenv
pydantic
sendgrid
openai-agents
